@@ -1,32 +1,32 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Dapr.Client" Version="1.12.0"/>
-    <PackageReference Include="Dapr.AspNetCore" Version="1.12.0"/>
-    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.13"/>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0"/>
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.0"/>
-    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.25.1"/>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="7.0.13"/>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="7.0.13"/>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.13"/>
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="7.0.3"/>
-    <PackageReference Include="amorphie.core" Version="1.1.17"/>
-    <PackageReference Include="amorphie.core.security" Version="1.0.0"/>
-    <PackageReference Include="IdentityModel.AspNetCore.OAuth2Introspection" Version="6.2.0"/>
-    <PackageReference Include="Microsoft.AspNetCore.SignalR.Core" Version="1.1.0"/>
-    <PackageReference Include="Microsoft.AspNetCore.SignalR.Protocols.MessagePack" Version="7.0.5"/>
-    <PackageReference Include="Microsoft.AspNetCore.SignalR.StackExchangeRedis" Version="7.0.5"/>
-    <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="7.0.5"/>
+    <PackageReference Include="Dapr.Client" Version="1.12.0" />
+    <PackageReference Include="Dapr.AspNetCore" Version="1.12.0" />
+    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.1" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.1" />
+    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="7.3.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="8.0.1" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="8.0.1" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.1" />
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.0" />
+    <PackageReference Include="amorphie.core" Version="1.1.25" />
+    <PackageReference Include="amorphie.core.security" Version="1.0.1" />
+    <PackageReference Include="IdentityModel.AspNetCore.OAuth2Introspection" Version="6.2.0" />
+    <PackageReference Include="Microsoft.AspNetCore.SignalR.Core" Version="1.1.0" />
+    <PackageReference Include="Microsoft.AspNetCore.SignalR.Protocols.MessagePack" Version="8.0.1" />
+    <PackageReference Include="Microsoft.AspNetCore.SignalR.StackExchangeRedis" Version="8.0.1" />
+    <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="8.0.1" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\amorphie.workflow.data\amorphie.workflow.data.csproj"/>
+    <ProjectReference Include="..\amorphie.workflow.data\amorphie.workflow.data.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\amorphie.workflow.core\amorphie.workflow.core.csproj"/>
+    <ProjectReference Include="..\amorphie.workflow.core\amorphie.workflow.core.csproj" />
   </ItemGroup>
 </Project>
